ENH : trata colisão de balas do inimigo com o heroi. Imprime pontuação e tempo decorrido na tela.

+++ headers/config.c
@@ -20,4 +20,4 @@ const int BULLET_STEP_SIZE = 15;
 const int FIRST_ENEMY_OFFSET   = 0;
 const int ENEMY_SPAWN_INTERVAL = 3;
 const int ENEMY_STEP_SIZE      = 2;
-const int ENEMY_FIRE_INTERVAL  = 1;
+const double ENEMY_FIRE_INTERVAL  = 1;
